Refactor roles to reduce code duplication
I'm not sure I'm happy with this refactor and may go back to the
one-role-per-tool structure, because I like the way Ansible will display
the role name in the task output. It makes it easier to follow the
output.

But, I wanted to at least capture this structure in the commit history.